+++ include/geometry/buffer.h
@@ -1,5 +1,5 @@
 /**
- * \file buffer_2.h
+ * \file buffer.h
  * \author Zhihao Zhan (zhanzhihao_dt@163.com)
  * \brief 几何收缩外扩
  * \version 0.1
@@ -9,12 +9,12 @@
  *
  */
 
-#ifndef GEOMETRY_BUFFER_2_H
-#define GEOMETRY_BUFFER_2_H
+#ifndef GEOMETRY_BUFFER_H
+#define GEOMETRY_BUFFER_H
 
 #include "bg.h"
 
-class Buffer_2 {
+class Buffer {
   public:
   /**
    * \brief 收缩 Segment
@@ -29,7 +29,7 @@ class Buffer_2 {
    *
    * \param pg
    * \param out
-   * 
+   *
    * \return true
    * \return false
    */
@@ -37,12 +37,11 @@ class Buffer_2 {
 
   /**
    * \brief 多边形bbox
-   * 
-   * \param pg 
+   *
+   * \param pg
    * \param bbox min_x, max_x, min_y, max_y
    */
-  static void BBox(const Polygon& pg, std::vector<double>& bbox);
-
+  static void BBox(const Polygon &pg, std::vector<double> &bbox);
 };
 
-#endif // GEOMETRY_BUFFER_2_H
+#endif // GEOMETRY_BUFFER_H