+++ test/geometry/test_buffer.cpp
@@ -1,22 +1,22 @@
 /**
- * \file test_buffer_2.cpp
+ * \file test_buffer.cpp
  * \author Zhihao Zhan (zhanzhihao_dt@163.com)
  * \brief 2d 几何外扩收缩以及bbox等测试
  * \version 0.1
  * \date 2023-06-12
- * 
+ *
  * @copyright Copyright (c) 2023
- * 
+ *
  */
 
-#include "geometry/utility_2.h"
-#include "visualization/vis_2.h"
+#include "geometry/utility.h"
+#include "visualization/vis.h"
 #include <iostream>
 
 int main()
 {
-  std::shared_ptr<Vis_2> vis_ptr = std::make_shared<Vis_2>();
-  
+  std::shared_ptr<Vis> vis_ptr = std::make_shared<Vis>();
+
   std::cout << " ================== 1. Shrink ================== " << std::endl;
   Point pt_before(0, 0);
   Point center(1, 1);
@@ -39,27 +39,27 @@ int main()
   bg::correct(pg2);
 
   Segment seg_before_shrink(pt_before, center);
-  Segment seg_after_shrink = Buffer_2::Shrink(seg_before_shrink);
-  std::cout << String_2::SegmentToString(seg_before_shrink) << " -> " << String_2::SegmentToString(seg_after_shrink)
+  Segment seg_after_shrink = Buffer::Shrink(seg_before_shrink);
+  std::cout << String::SegmentToString(seg_before_shrink) << " -> " << String::SegmentToString(seg_after_shrink)
             << std::endl;
 
   Segment seg_before_shrink1(center, pt_before);
-  Segment seg_after_shrink1 = Buffer_2::Shrink(seg_before_shrink1);
-  std::cout << String_2::SegmentToString(seg_before_shrink1) << " -> " << String_2::SegmentToString(seg_after_shrink1)
+  Segment seg_after_shrink1 = Buffer::Shrink(seg_before_shrink1);
+  std::cout << String::SegmentToString(seg_before_shrink1) << " -> " << String::SegmentToString(seg_after_shrink1)
             << std::endl;
 
   Segment seg_before_shrink2(pt_before, Point(0, 2));
-  Segment seg_after_shrink2 = Buffer_2::Shrink(seg_before_shrink2);
-  std::cout << String_2::SegmentToString(seg_before_shrink2) << " -> " << String_2::SegmentToString(seg_after_shrink2)
+  Segment seg_after_shrink2 = Buffer::Shrink(seg_before_shrink2);
+  std::cout << String::SegmentToString(seg_before_shrink2) << " -> " << String::SegmentToString(seg_after_shrink2)
             << std::endl;
 
   Segment seg_before_shrink3(pt_before, Point(2, 0));
-  Segment seg_after_shrink3 = Buffer_2::Shrink(seg_before_shrink3);
-  std::cout << String_2::SegmentToString(seg_before_shrink3) << " -> " << String_2::SegmentToString(seg_after_shrink3)
+  Segment seg_after_shrink3 = Buffer::Shrink(seg_before_shrink3);
+  std::cout << String::SegmentToString(seg_before_shrink3) << " -> " << String::SegmentToString(seg_after_shrink3)
             << std::endl;
 
   Polygon pg1_after_shrink;
-  Buffer_2::Shrink(pg1, pg1_after_shrink);
+  Buffer::Shrink(pg1, pg1_after_shrink);
   if (0)
   {
     vis_ptr->Clear();
@@ -71,7 +71,7 @@ int main()
   }
 
   Polygon pg2_after_shrink;
-  Buffer_2::Shrink(pg2, pg2_after_shrink);
+  Buffer::Shrink(pg2, pg2_after_shrink);
   if (1)
   {
     vis_ptr->Clear();